Replace the REST version of the API with new sdk in all places
everything works
