+++ core/aimodels/gemini_rest/__init__.py
