+++ core/aimodels/gemini_rest/helper_methods.py
